+++ internal/types/resourceGroups.go
@@ -1,6 +1,8 @@
 package types
 
-import pkgtypes "github.com/konstructio/kubefirst-api/pkg/types"
+import (
+	pkgtypes "github.com/konstructio/kubefirst-api/pkg/types"
+)
 
 type ResourceGroupsListRequest struct {
 	AzureAuth pkgtypes.AzureAuth `bson:"azure_auth,omitempty" json:"azure_auth,omitempty"`
